Found issue while testing global config settings
The global configuration change made dropdowns not populate as they were
no longer part of that build step config.  This fixes that.
